Version 1.1: UI Improvements
Hello! Welcome to the newest version of the Simple Texteditor!

This version contains some UI improvements for better coding. This include a new Dark Mode and better colors for objects. The themes are also optimized.

There is also a new question mark with information about the editor, and how to create code for Arya ( https://ryanvanderhaak.github.io/Online-Leerhulp/Arya.html ).

Happy coding and see you in the next update!
Ryan